+++ tests/testthat/unit/test-filter_on.R
@@ -1,11 +1,5 @@
 context("  Filter on")
 
-test_that("An unnamed value results in an error.", {
-    expect_error({
-        DT %>% start_expr %>% filter_on("foo") %>% end_expr
-    })
-})
-
 test_that("The filter_on verb works as expected.", {
     ans <- DT %>% start_expr %>% filter_on(cyl = 6, am = 0) %>% end_expr
     expect_identical(ans, DT[.(6, 0), on = c("cyl", "am")])
@@ -15,6 +9,20 @@ test_that("The filter_on verb works as expected.", {
 
     ans <- DT %>% start_expr %>% filter_on(cyl = 6, am = 0, mult = "last") %>% end_expr
     expect_identical(ans, DT[.(6, 0), on = c("cyl", "am"), mult = "last"])
+
+    # ----------------------------------------------------------------------------------------------
+
+    expected <- data.table::setkey(data.table::copy(DT), cyl, am)[.(6, 0)]
+
+    ans <- DT %>%
+        data.table::copy(.) %>%
+        data.table::setkey(cyl, am) %>%
+        start_expr %>% filter_on(cyl = 6, 0) %>%
+        end_expr
+
+    expect_identical(ans, expected)
+
+    # ----------------------------------------------------------------------------------------------
 
     ans <- DT %>% start_expr %>% filter_on(cyl = 10) %>% end_expr
     expect_identical(nrow(ans), 1L)